[REFACTOR] 폴더 이름 변경 및 파일 위치 변경
