+++ app/src/main/java/com/example/leaveit/MainActivity.kt
@@ -4,7 +4,7 @@ import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import androidx.fragment.app.Fragment
 import com.example.leaveit.databinding.ActivityMainBinding
-import com.example.leaveit.presentation.review.makereview.SelectPlaceForReview
+import com.example.leaveit.presentation.review.SelectPlaceForReview
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
