--- app/src/main/java/com/example/leaveit/presentation/review/makereview/MakeReviewView.kt
@@ -1,4 +1,0 @@
-package com.example.leaveit.presentation.review.makereview
-
-class MakeReviewView {
-}
