+++ app/src/main/java/com/example/leaveit/presentation/review/postview/MakeReviewView.kt
@@ -1,0 +1,5 @@
+package com.example.leaveit.presentation.review.postview
+
+class MakeReviewView {
+
+}
